Multi-input socket and notify
Sockets can now be set up to accept multiple inputs (such as you might
want on an add function block) or a single input (such as you want on a
scalar sink). To inform the user that the connection cannot be made,
I've added a generic notify module. The Information box pops up behind
the blocks but I sort of like that.