Properties Views use Prototypical inheritance
With hindsight, using composition for the object hierarchy required a
load of boilerplate that obscured the use of the views. I have changed
PropertiesViewBase to be more like a base class (using prototypical
inheritance). A good example of where this is brilliant is in the scalar
sink view, which is now just a few lines.