Harness input connection to configuration
It is now possible to connect an input socket with a configuration item.
Also fixed a strange problem with the brute force engine not resetting
input ready states and the array source showing the next pattern rather
than the last.

Some infrastructure added for the next piece, which is connecting
outputs to configuration items.